Swap Tinyshee and Demo Player
It makes more sense for the demo which uses Banshee.Services
to be called Tinyshee (Tiny + Banshee), whereas the client
which uses the PlayerEngine backend directly is more of a
proof-of-concept. Therefore swap the names.

This also includes minor tweaks to layout and commenting,
as well as changes the window title of each demo to match.