Modernise Build System (#4)
* Use Hyena.Addins nuget and drop Mono.Addins
* Update Hyena submodule
* Update build instructions

Simplify the building process by dropping the git
submodule for mono-addins and replacing it with Hyena.Addins,
a mono-addins fork for netstandard2.1. We now use a test
nuget instead of a submodule, which is obtained automatically
by MSBuild.

Additionally use a newer version of Hyena, and update build
instructions to reflect the new, easier procedure for building.